criando itens pedidos

+++ src/main/java/app/projetaria/appcommerce/entity/Produto.java
@@ -1,7 +1,5 @@
 package app.projetaria.appcommerce.entity;
 
-import java.math.BigDecimal;
-
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -18,7 +16,7 @@ public class Produto {
 	
 	private String descricao;
 	
-	private BigDecimal preco;
+	private Double preco;
 
 	public Integer getId() {
 		return id;
@@ -44,11 +42,11 @@ public class Produto {
 		this.descricao = descricao;
 	}
 
-	public BigDecimal getPreco() {
+	public Double getPreco() {
 		return preco;
 	}
 
-	public void setPreco(BigDecimal preco) {
+	public void setPreco(Double preco) {
 		this.preco = preco;
 	}
